#2081: only remove the performed skill when a passed quiz attempt is being removed

+++ service/src/main/java/skills/services/quiz/QuizDefService.groovy
@@ -437,13 +437,15 @@ class QuizDefService {
 
         userQuizAttemptRepo.delete(userQuizAttempt)
         log.info("Removed [{}]", userQuizAttempt.toString())
-        List<QuizToSkillDefRepo.ProjectIdAndSkillId> linkedSkills = quizToSkillDefRepo.getSkillsForQuiz(quizDef1.id)
-        linkedSkills.each {
-            List<UserPerformedSkill> userPerformedSkills = userPerformedSkillRepo.findAllBySkillRefIdAndUserId(it.skillRefId, userQuizAttempt.userId)
-            if (userPerformedSkills) {
-                // since this is a quiz-based skill should never be more than 1
-                userPerformedSkills.each {
-                    skillEventAdminService.removePerformedSkillEvent(it)
+        if (userQuizAttempt.status == UserQuizAttempt.QuizAttemptStatus.PASSED) {
+            List<QuizToSkillDefRepo.ProjectIdAndSkillId> linkedSkills = quizToSkillDefRepo.getSkillsForQuiz(quizDef1.id)
+            linkedSkills.each {
+                List<UserPerformedSkill> userPerformedSkills = userPerformedSkillRepo.findAllBySkillRefIdAndUserId(it.skillRefId, userQuizAttempt.userId)
+                if (userPerformedSkills) {
+                    // since this is a quiz-based skill should never be more than 1
+                    userPerformedSkills.each {
+                        skillEventAdminService.removePerformedSkillEvent(it)
+                    }
                 }
             }
         }
